Worked on models/user_model
Ah, but we do want the q-marks; they result in the db engine escaping the
values for us.

I reindented the whole thing, I'm neurotic like that, and documented the
existing functions.  I also finishted the modify_user function and changed its calling
semantics somewhat so that it can be used for updates that both do and do not
change the user's password.

The salt_password function was worked into the process where needed.